+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 5.14.0 - 2025-10-21
 
 ### Added
+- [#304](https://github.com/deviceinsight/kafkactl/pull/304) Add support for headers when using file input to produce 
 - [#300](https://github.com/deviceinsight/kafkactl/issues/300) Allow mounting secrets with certificates to kafkactl pod
 - [#253](https://github.com/deviceinsight/kafkactl/issues/253) Allow configuring k8s resource requests/limits